+++ src/test/java/org/tron/core/db2/SnapshotManagerTest.java
@@ -15,6 +15,7 @@ import org.tron.core.config.DefaultConfig;
 import org.tron.core.config.args.Args;
 import org.tron.core.db2.RevokingDbWithCacheNewValueTest.TestRevokingTronStore;
 import org.tron.core.db2.RevokingDbWithCacheNewValueTest.TestSnapshotManager;
+import org.tron.core.db2.SnapshotRootTest.ProtoCapsuleTest;
 import org.tron.core.db2.core.ISession;
 import org.tron.core.db2.core.SnapshotManager;
 import org.tron.core.exception.BadItemException;
@@ -95,7 +96,7 @@ public class SnapshotManagerTest {
       }
     }
     Assert.assertEquals(null,
-      tronDatabase.get(protoCapsule.getData()));
+        tronDatabase.get(protoCapsule.getData()));
 
   }
